examples

+++ bin/popmark.dart
@@ -15,6 +15,8 @@ Future<void> main(List<String> arguments) async {
         return (ArgParser()
               ..addOption('imports', abbr: 'i')
               ..addOption('output', abbr: 'o', defaultsTo: targetFile)
+              ..addOption('template', abbr: 't', defaultsTo: 'DEFAULT')
+              ..addOption('execute', abbr: 'e', defaultsTo: 'ALL')
               ..addFlag('help', abbr: 'h', defaultsTo: false)
               ..addFlag('cleanup', abbr: 'c', defaultsTo: true)
               ..addFlag('strip', abbr: 's', defaultsTo: false))
@@ -26,8 +28,46 @@ Future<void> main(List<String> arguments) async {
       return null;
     }(),
 
-        /// The name of the file to write to.
+        // The name of the file to write to.
         output = results['output'] as String,
+
+        // The template to use for temporary Dart files.
+        template = (results['template'] as String).toUpperCase() == 'DEFAULT'
+            ? defaultTemplate
+            : await () async {
+                final templateFile = results['template'] as String;
+                if (await File(templateFile).exists()) {
+                  return await File(templateFile).readAsString();
+                } else {
+                  print(
+                      'Cannot find file "$templateFile". For help, run: popmark --help');
+                  exit(0);
+                }
+                return null;
+              }(),
+
+        // Which code segments to / not to execute.
+        doExecute = -10,
+        doNotExecute = -20,
+        executeSegments = () {
+      final executeOption = results['execute'] as String;
+      if (executeOption.toUpperCase() == 'ALL') {
+        return null;
+      } else {
+        final cleanExecuteOptions =
+            executeOption.replaceAll(RegExp(r'[^*,;0-9]'), '');
+
+        return <int>[
+          cleanExecuteOptions.isNotEmpty && cleanExecuteOptions[0] == '*'
+              ? doNotExecute
+              : doExecute,
+          ...cleanExecuteOptions
+              .replaceAll('*', '')
+              .split(RegExp(r'[,;]'))
+              .map((x) => int.tryParse(x).abs() - 1)
+        ];
+      }
+    }(),
 
         // Whether to delete the temporary Dart files afterwards.
         cleanup = results['cleanup'] as bool,
@@ -84,10 +124,12 @@ Future<void> main(List<String> arguments) async {
     var
         // The state of the reader, which determines response to lines.
         state = State.markdown,
+
         // A buffer for the code to be executed.
         code = StringBuffer(),
+
         // An index tracker for the code segments.
-        index = 0;
+        segmentIndex = 0;
 
     final contentBuffer = StringBuffer();
     for (final line in lines) {
@@ -109,40 +151,46 @@ Future<void> main(List<String> arguments) async {
 
           if (line.contains(closeCode)) {
             // Execute the code segment and insert its output.
-            final dartFile = '_temp_popmark$index.dart';
-            index++;
+            if (executeSegments == null ||
+                (executeSegments.first == doExecute &&
+                    executeSegments.contains(segmentIndex)) ||
+                (executeSegments.first == doNotExecute &&
+                    !executeSegments.contains(segmentIndex))) {
+              final dartFile = '_temp_popmark$segmentIndex.dart';
 
-            await File(dartFile).writeAsString(template
-                .replaceFirst('{IMPORTS}', importLines.join('\n'))
-                .replaceFirst('{BODY}', code.toString()));
+              await File(dartFile).writeAsString(template
+                  .replaceFirst('{IMPORTS}', importLines.join('\n'))
+                  .replaceFirst('{BODY}', code.toString()));
 
-            final result = await Process.run('dart', [dartFile]);
+              final result = await Process.run('dart', [dartFile]);
 
-            if (!strip) {
-              final resultStdout = result.stdout.toString(),
-                  output = resultStdout.isNotEmpty
-                      ? resultStdout
-                      : result.stderr
-                          .toString()
-                          .split('\n')
-                          .map((line) => line.contains('_temp_popmark')
-                              ? line.split(' ').sublist(1).join(' ')
-                              : line)
-                          .join('\n');
+              if (!strip) {
+                final resultStdout = result.stdout.toString(),
+                    output = resultStdout.isNotEmpty
+                        ? resultStdout
+                        : result.stderr
+                            .toString()
+                            .split('\n')
+                            .map((line) => line.contains('_temp_popmark')
+                                ? line.split(' ').sublist(1).join(' ')
+                                : line)
+                            .join('\n');
 
-              contentBuffer.writeln('\n$openPopmark\n$output$closePopmark\n');
-            }
+                contentBuffer.writeln('\n$openPopmark\n$output$closePopmark\n');
+              }
 
-            if (cleanup) {
-              await File(dartFile).delete();
-            } else {
-              final cleanCode =
-                  (await Process.run('dartfmt', [dartFile])).stdout as String;
-              await File(dartFile).writeAsString(cleanCode);
+              if (cleanup) {
+                await File(dartFile).delete();
+              } else {
+                final cleanCode =
+                    (await Process.run('dartfmt', [dartFile])).stdout as String;
+                await File(dartFile).writeAsString(cleanCode);
+              }
             }
 
             code.clear();
             state = State.markdown;
+            segmentIndex++;
           } else {
             // Include code segments.
             code.writeln(line);
@@ -185,51 +233,65 @@ Hello, world!
 
 Basic use:
 
-  pub run popmark [file] [options] [flags]
+  popmark [file] [options] [flags]
 
 Options:
 
-  --help      -h    Shows this help.
+  --execute   -e    Identifies which code segments to or not to execute.
+                    For example, to only execute the 1st and the 3rd
+                    code segment, use:
 
-                    Example:
-                    pub run popmark -h
+                    popmark target.md -e 1,3
 
-  --output    -o    Specifies the file to write to. If not set, popmark 
-                    writes to the target file.
+                    To execute all segments except for the 1st and 3rd
+                    segment, use an asterisk:
 
-                    Example:
-                    pub run popmark target.md -o out.md
+                    popmark target.md -e *1,3
 
-  --imports   -i    Specifies any libraries or packages the documented 
-                    code relies on, separated by semi-colons.
+  --help      -h    To get help (i.e. see this information), use:
 
-                    Example:
-                    pub run popmark target.md -i 'dart:math'
+                    popmark -h
+
+  --output    -o    By default, popmark writes to the target file. To 
+                    specify the file to write to, set the output file:
+
+                    popmark target.md -o out.md
+
+  --imports   -i    Specify any libraries or packages the documented 
+                    code relies on, separated by semi-colons:
+
+                    popmark target.md -i 'dart:io;dart:math'
+
+  --template  -t    Specify the path to the template Dart code to use.
+                    For example, if template.txt contains Dart code with
+                    the text {BODY} to indicate where the documented code
+                    should be inserted, we could use it as a template:
+
+                    popmark target.md -t template.txt
 
 Flags:
 
-  --cleanup   -c    Specifies whether to delete the Dart files run in 
-                    the background (popmark cleans up by default).
+  --cleanup   -c    By default, popmark cleans up after itself. Specify 
+                    whether to delete the Dart files popmark runs in the 
+                    background using:
 
-                    Example:
-                    pub run popmark target.md --no-cleanup
+                    popmark target.md --no-cleanup
 
-  --strip     -s    Specifies whether to remove the program segment output.
+  --strip     -s    Strip all code segment output:
 
-                    Example:
-                    pub run popmark target.md -s
+                    popmark target.md -s
 
 For more information, check out the wiki at ...
 
 ''');
 }
 
-final template = '''
+final defaultTemplate = '''
 // Generated by popmark.
 
 {IMPORTS}
 
-void main() {
+main() async {
 {BODY}
 }
 ''';
